mm: change flags' type from u64 to usize

+++ kernel/mm/types.rs
@@ -100,11 +100,11 @@ pub trait RootPageDirOps {
     fn switch_to_this(&self);
     fn walk_dir(&mut self, addr: VirtAddr, create: bool) -> Option<&mut LeafDirEntry>;
     fn walk_dir_large(&mut self, addr: VirtAddr, create: bool) -> Option<&mut LeafDirEntryLarge>;
-    fn map_page_at_addr(&mut self, page: &mut PageInfo, addr: VirtAddr, perms: u64);
+    fn map_page_at_addr(&mut self, page: &mut PageInfo, addr: VirtAddr, perms: usize);
     fn unmap_page_at_addr(&mut self, addr: VirtAddr);
-    fn map_static_region(&mut self, from: VirtAddr, to: PhysAddr, npages: usize, perms: u64);
+    fn map_static_region(&mut self, from: VirtAddr, to: PhysAddr, npages: usize, perms: usize);
 
-    fn alloc_range(&mut self, addr: VirtAddr, size: usize, perms: u64) {
+    fn alloc_range(&mut self, addr: VirtAddr, size: usize, perms: usize) {
         let beg = addr.0.page_round_down();
         let end = addr.0.checked_add(size).unwrap().page_round_up();
